vyos/tests/requirements.txt
cloudshell-snmp>=4.0.2,<5.0.0
cloudshell-cli>=4.0.4,<5.0.0